refactor(memory): split default context helpers [impact-checked]
