+++ dharma_swarm/context_compiler.py
@@ -7,13 +7,9 @@ from pathlib import Path
 from typing import Any, Callable
 
 from dharma_swarm.context_compiler_utils import (
-    ContextSection,
-    approx_char_budget as _approx_char_budget,
-    canonical_json as _canonical_json,
-    context_scan_metadata as _context_scan_metadata,
-    dedupe_strings as _dedupe,
-    sha256_text as _sha256,
-    truncate_text as _truncate,
+    ContextSection, approx_char_budget as _approx_char_budget,
+    canonical_json as _canonical_json, context_scan_metadata as _context_scan_metadata,
+    dedupe_strings as _dedupe, sha256_text as _sha256, truncate_text as _truncate,
     utc_now as _utc_now,
 )
 from dharma_swarm.memory_lattice import MemoryLattice, MemoryRecallHit
@@ -25,22 +21,11 @@ from dharma_swarm.memory_kernel.context_compiler_shadow import (
     notify_memory_kernel_context_canary,
     run_memory_kernel_context_canary,
 )
-from dharma_swarm.memory_kernel import (
-    MemoryContextBudget,
-    MemoryContextPack,
-    MemoryKernel,
-    MemoryQuery,
-    TruthState,
-)
+from dharma_swarm.memory_kernel.default_context import build_memory_kernel_default_context
 from dharma_swarm.provider_policy import ProviderPolicyRouter, ProviderRouteRequest
 from dharma_swarm.runtime_state import (
-    ArtifactRecord,
-    ContextBundleRecord,
-    DelegationRun,
-    MemoryFact,
-    RuntimeStateStore,
-    SessionState,
-    WorkspaceLease,
+    ArtifactRecord, ContextBundleRecord, DelegationRun, MemoryFact, RuntimeStateStore,
+    SessionState, WorkspaceLease,
 )
 
 logger = logging.getLogger(__name__)
@@ -81,7 +66,7 @@ class ContextCompiler:
         memory_palace: Any = None,
         graph_store: Any = None,
         knowledge_store: Any = None,
-        memory_kernel: MemoryKernel | None = None,
+        memory_kernel: Any = None,
     ) -> None:
         self.runtime_state = runtime_state
         self.memory_lattice = memory_lattice
@@ -136,9 +121,7 @@ class ContextCompiler:
         memory_kernel_shadow_surfaces: tuple[str, ...] = (),
         memory_kernel_shadow_callback: Callable[[object], object] | None = None,
     ) -> ContextBundleRecord:
-        # ── MemPO-style truncation ──────────────────────────────────
-        # When ENABLE_MEM_TRUNCATION is set and a previous_mem is provided,
-        # replace the full context assembly with a compact summary.
+        # Optional MemPO-style truncation can replace full assembly.
         import os as _os
         _mem_truncation = _os.getenv("ENABLE_MEM_TRUNCATION", "false").strip().lower()
         if (
@@ -289,9 +272,8 @@ class ContextCompiler:
             query=query,
             task_id=task_id,
         )
-        memory_kernel_section, memory_kernel_metadata = self._memory_kernel_section(
-            recall_query=recall_query,
-            token_budget=token_budget,
+        memory_kernel_section, memory_kernel_metadata = build_memory_kernel_default_context(
+            self.memory_kernel, recall_query=recall_query, token_budget=token_budget
         )
         recall_hits = (
             await self.memory_lattice.recall(
@@ -391,9 +373,7 @@ class ContextCompiler:
                     metadata=dict(canary.metadata),
                 ),
             ]
-            source_refs = _dedupe(
-                (*source_refs, *memory_kernel_context_source_refs(canary))
-            )
+            source_refs = _dedupe((*source_refs, *memory_kernel_context_source_refs(canary)))
         created_at = _utc_now()
         checksum = _sha256(
             _canonical_json(
@@ -552,7 +532,6 @@ class ContextCompiler:
 
         if memory_kernel_section is not None:
             sections.append(memory_kernel_section)
-
         if always_on.strip():
             sections.append(
                 ContextSection(
@@ -712,69 +691,6 @@ class ContextCompiler:
             )
 
         return sections
-
-    def _memory_kernel_section(
-        self,
-        *,
-        recall_query: str,
-        token_budget: int,
-    ) -> tuple[ContextSection | None, dict[str, Any]]:
-        if self.memory_kernel is None:
-            return None, {"status": "not_configured"}
-
-        try:
-            atom_budget = max(4, min(24, int(token_budget) // 100))
-            pack = self.memory_kernel.preview_memory_pack(
-                query=MemoryQuery(
-                    limit_total=atom_budget,
-                    limit_per_surface=atom_budget,
-                    include_content=True,
-                    max_canon_risk=None,
-                    max_pii_risk=None,
-                    include_high_risk=False,
-                    include_projections=False,
-                    include_unsafe=False,
-                    require_source_digest=True,
-                    require_source_row_key=True,
-                ),
-                budget=MemoryContextBudget(
-                    max_candidate_atoms=atom_budget,
-                    max_admitted_atoms=max(1, min(8, atom_budget)),
-                    max_total_chars=max(600, min(2400, int(token_budget) * 2)),
-                    max_atom_chars=420,
-                    include_content=True,
-                    require_context_admissible=False,
-                    allow_projections=False,
-                    allow_high_risk=False,
-                    allowed_truth_states=(
-                        TruthState.OBSERVED,
-                        TruthState.CLAIMED,
-                        TruthState.CURATED,
-                        TruthState.CANONICAL,
-                    ),
-                ),
-            )
-        except Exception as exc:
-            logger.debug("Memory Kernel default context failed", exc_info=True)
-            return None, {
-                "status": "failed",
-                "error_type": type(exc).__name__,
-                "error": str(exc)[:240],
-            }
-
-        metadata = _memory_kernel_pack_metadata(pack)
-        metadata["status"] = "used"
-        metadata["query_present"] = bool(recall_query)
-        return (
-            ContextSection(
-                name="Memory Kernel",
-                priority=4,
-                content=_format_memory_kernel_pack(pack),
-                source_refs=_memory_kernel_pack_source_refs(pack),
-                metadata=metadata,
-            ),
-            metadata,
-        )
 
     def _fit_sections(
         self,
@@ -1080,75 +996,3 @@ class ContextCompiler:
             logger.debug("Semantic graph query failed: %s", exc)
 
         return results
-
-
-def _memory_kernel_pack_metadata(pack: MemoryContextPack) -> dict[str, Any]:
-    return {
-        "pack_id": pack.pack_id,
-        "candidate_count": pack.candidate_count,
-        "admitted_count": pack.admitted_count,
-        "omitted_count": pack.omitted_count,
-        "candidate_truncated": pack.candidate_truncated,
-        "warnings": list(pack.warnings),
-    }
-
-
-def _memory_kernel_pack_source_refs(pack: MemoryContextPack) -> list[str]:
-    refs: list[str] = []
-    for item in pack.items:
-        if not item.admitted:
-            continue
-        refs.append(f"memory_kernel:{item.surface_id}")
-        refs.extend(item.source_refs)
-    return _dedupe(refs)
-
-
-def _format_memory_kernel_pack(pack: MemoryContextPack) -> str:
-    lines = [
-        f"- pack_id={pack.pack_id}",
-        f"- candidates={pack.candidate_count}",
-        f"- admitted={pack.admitted_count}",
-        f"- omitted={pack.omitted_count}",
-    ]
-    if pack.warnings:
-        lines.append(f"- warnings={', '.join(str(item) for item in pack.warnings)}")
-
-    admitted = [item for item in pack.items if item.admitted]
-    if admitted:
-        lines.extend(["", "Admitted atoms:"])
-        for item in admitted[:8]:
-            lines.append(
-                "- "
-                f"rank={item.rank} surface={item.surface_id} "
-                f"type={_enum_value(item.atom_type)} truth={_enum_value(item.truth_state)} "
-                f"authority={_enum_value(item.authority_level)}"
-            )
-            if item.selection_reasons:
-                lines.append(f"  reasons={', '.join(item.selection_reasons)}")
-            if item.content_snippet:
-                lines.append(f"  snippet={_inline_context(item.content_snippet, 420)}")
-    else:
-        lines.extend(["", "Admitted atoms:", "- none"])
-
-    omitted = [item for item in pack.items if not item.admitted]
-    if omitted:
-        lines.extend(["", "Omitted atoms:"])
-        for item in omitted[:8]:
-            lines.append(
-                "- "
-                f"surface={item.surface_id} type={_enum_value(item.atom_type)} "
-                f"truth={_enum_value(item.truth_state)} "
-                f"reasons={', '.join(item.omission_reasons)}"
-            )
-    return "\n".join(lines)
-
-
-def _enum_value(value: Any) -> str:
-    return str(getattr(value, "value", value))
-
-
-def _inline_context(value: str, max_chars: int) -> str:
-    text = " ".join(str(value).split())
-    if len(text) <= max_chars:
-        return text
-    return text[: max(0, max_chars - 14)].rstrip() + "...<truncated>"
